Fixing a pointer cast for 64-bits archs

+++ neural_cluster.c
@@ -117,17 +117,17 @@ __AI_neural_clusters_to_xml ( kmeans_t *km, AI_alerts_per_neuron *alerts_per_neu
 							{
 								/* Yeah, dirty horrible way for avoiding some SEGFAULTs for invalid memory access,
 								 * but it's the only working way on some 64-bits architectures */
-								if ( km->clusters < 0x100 )
+								if ((unsigned long int) km->clusters < 0x100 )
 								{
 									continue;
 								}
 
-								if ( km->clusters[l] < 0x100 )
+								if ((unsigned long int) km->clusters[l] < 0x100 )
 								{
 									continue;
 								}
 
-								if ( km->clusters[l][m] < 0x100 )
+								if ((unsigned long int) km->clusters[l][m] < 0x100 )
 								{
 									continue;
 								}
